app/logic/app/requirements.txt
flask==1.0.2
Flask-Cors==1.10.2
numpy==1.17.3
scipy==1.3.1
matplotlib==3.1.1
Pillow==6.2.1
